requirements.txt
Pillow==2.7.0
beautifulsoup4==4.3.2
numpy==1.8.2
opencv-python==2.4.10
pygame==1.9.2a0
python-tesseract==0.9